Cast to float, check size of input data

+++ io_utils.py
@@ -8,8 +8,8 @@ def read_data(input_file: str) -> Tuple[numpy.ndarray, numpy.ndarray]:
     with open(input_file) as f:
         for line in f:
             x, y = line.split()
-            xs.append(int(x))
-            ys.append(int(y))
+            xs.append(float(x))
+            ys.append(float(y))
     xs = numpy.array(xs)
     ys = numpy.array(ys)
     sorted_idx = numpy.argsort(xs)
